Need only one cast here

+++ ImageComputation.cpp
@@ -23,16 +23,16 @@ double averagePixelDistance(Image& a, Image& b) {
 			const PIXEL* pixelA = a.get(x, y);
 			const PIXEL* pixelB = b.get(x, y);
 
-			int diffR = (int) pixelA->r - (int) pixelB->r;
+			int diffR = (int) pixelA->r - pixelB->r;
 			diffR *= diffR;
 
-			int diffG = (int) pixelA->g - (int) pixelB->g;
+			int diffG = (int) pixelA->g - pixelB->g;
 			diffG *= diffG;
 
-			int diffB = (int) pixelA->b - (int) pixelB->b;
+			int diffB = (int) pixelA->b - pixelB->b;
 			diffB *= diffB;
 
-			int diffA = (int) pixelA->a - (int) pixelB->a;
+			int diffA = (int) pixelA->a - pixelB->a;
 			diffA *= diffA;
 
 			sumDistance += sqrt(diffR + diffG + diffB + diffA);
